+++ backend/src/routes.js
@@ -1,18 +1,9 @@
 const express = require("express");
-const crypto = require("crypto");
-const connection = require("./database/connection");
+const ongController = require("./controllers/ongController")
 const routes = express.Router();
 
-routes.get("/ongs", async (request, response) => {
-  const ongs = await connection("ongs").select("*");
-  return response.json(ongs);
-});
-
-routes.post("/ongs", async (request, response) => {
-  const { name, email, cel, city, uf } = request.body;
-  const id = crypto.randomBytes(4).toString("HEX");
-  await connection("ongs").insert({ id, name, email, cel, city, uf });
-  return response.json({ id });
-});
+routes.get("/ongs", ongController.selectAll );
+routes.get("/ongs:name", ongController.selectByName );
+routes.post("/ongs", ongController.create );
 
 module.exports = routes;
